backend/requirements.txt
bs4
langchain
langchain-community
langchain-openai
langchain-chroma
requests
langgraph
black
langserve
sse_starlette
fastapi
uvicorn
fastapi-cors
youtube-transcript-api
python-dotenv
pydantic
langchain-core
langchain-text-splitters
pytube